forms->plist - rewrite using loop
Matter of preference, I guess. I actually like the `cond`, it's more
familiar to me than the behavior of loop's conditionals. But the
lambda, extra indentation, and the list being operated on being so far
away from the start of the operation (reminds me of Mark Twain talking
about separable verbs in German :>), and close to what is being
returned, such that one is likely to confuse one with the other...not
so much. :D